Add files via upload
An Excel based game to enhance the understanding of the Mode of Action for Antibiotics. This was designed and run over a virtual environment for second year (Level 5, UK system) students. This game is linked to a practical session on Mode of Action (using chloramphenicol, ampicillin and streptomycin on E. coli).

Some student comments:
"The games and interactivity is super helpful for different learning styles."
"The [most enjoyable part of the unit were the] activities whilst on the call (antibiotic game!!!!!!)"