PowerPoint based biochemical adventure game
This PowerPoint adventure game, was originally designed for first year (Level 4, UK system) undergraduates learning about Animal Cell structure and Cell Cycle as an virtual session through Teams. The idea was that the students could use the functionality in Teams to choose their own path through the Beta-cell to make insulin or divide, learning about checkpoints and the sturcture and uses of the organelles within the game.