Rules for an Excel based antibiotics game
These files are the rules and initial phenotype card for a First year/Level 4 (UK system) undergraduate microbiology course, looking at antibiotics and how they work. This was originally developed to be use in a virtual session coupled with videos on antibiotic action.

This probably requires tweaking as many students (~60%) "died" within the first round.